Add Gunicorn using gthreads as wsgi server
If you want to run Swift without eventlet you need to set the
environment variable "USE_EVENTLET=0". For example, using swift-init:

        USE_EVENTLET=0 swift-init restart main

You can also uninstall eventlet before (re-)starting the Swift services
to ensure it is working without it.

Due to a more strict validation of requests, headers and different
encoding a bit of monkey patching is required to match the current
behaviour running Swift with eventlet.

To tune gunicorns a new setting "threads" has been added to the
*-server.conf files. It defaults to 4, but especially the proxy-server
and object-server might be more performant using a higher setting. Note
that this is the number of threads per worker; the number of workers is
by default set to the number of CPU-cores.

Functional tests (still requiring eventlet at this point) should pass
with this implementation, except for the xprofile middleware which
depends on eventlet.

This implementation does not yet work with EC; this needs additional
changes to support 100-Continue requests and will be added in a
followup.

Change-Id: I9a90376a67d084e943446ba1c82c3be197187155
Signed-off-by: Christian Schwede <cschwede@mailbox.org>