Add replacement for eventlet.pools.Pool
Provides a simplified version of eventlet.pools (not implementing
functions unused in Swift) and implements a thread-safe drop-in
replacement for eventlet.pools.Pool using threading.Condition.

The context manager get() acquires a pool item, yields it to the
caller, and guarantees the item is returned via put() in a finally
block, preventing leaks if exceptions raised.

Change-Id: Ic54d7c930e3796024fd724683c397e48a5cb00b6
Signed-off-by: Christian Schwede <cschwede@mailbox.org>