Skip importing eventlet in threading mode
This makes eventlets imports optional if running in threading mode. Not
all imports have been made optional, this patch starts with the minimum
set of files to make it possible to run Swift API services using
gunicorn without importing eventlet.

swift.common.http_protocol.SwiftHttpProtocol/SwiftHttpProxiedProtocol
are not replaced because these are only used if eventlet is used.

This disables zero-copy in threading mode and needs a follow up to
implement this properly when using gunicorn.

Change-Id: I650d08d638479c9795e55bde7f8c58fd47af152c
Signed-off-by: Christian Schwede <cschwede@mailbox.org>